refactor(web): shared UserRoleAction + users route to directory form
Extract the grant/revoke wiring (confirm dialog, setUserRole action call,
own-row hiding, toasts) into components/admin/users/user-role-action.tsx —
the detail page is its second call site. Move users.tsx to users/index.tsx
(the messages/ dynamic-sibling convention) and link each member cell to
/admin/users/$userId. routeTree.gen.ts regenerates on the next vite dev.
